Most of the work to make the shared copy work with netfx

+++ src/Microsoft.Data.SqlClient/src/Interop/Windows/NtDll/Interop.SecurityQualityOfService.cs
@@ -14,7 +14,7 @@ internal partial class Interop
             bool isDynamicTracking,
             bool isEffectiveOnly)
         {
-            Length = sizeof(SecurityQualityOfService);
+            Length = (uint)sizeof(SecurityQualityOfService);
             ImpersonationLevel = impersonationLevel;
             IsDynamicTracking = isDynamicTracking;
             IsEffectiveOnly = isEffectiveOnly;
